Eigen 3DoF registration

+++ src/app/localization/radar_odometry_ros/include/algorithm/ego_motion_compensation.hpp
@@ -24,14 +24,14 @@
 namespace radar_odometry {
 
 pcl::PointCloud<PointXYZPRVAE>::Ptr VelFiltering(const pcl::PointCloud<PointXYZPRVAE>::Ptr& cloud, 
-                                                const Eigen::Matrix4f &predicted_vel,
+                                                const Eigen::Matrix4d &predicted_vel,
                                                 float margin);
 
 // Return coeff of y = c_0 * cos(x) + c_1 * sin(x)
-Eigen::Vector2f FitSine(const pcl::PointCloud<PointXYZPRVAE>::Ptr& cloud);
+Eigen::Vector2d FitSine(const pcl::PointCloud<PointXYZPRVAE>::Ptr& cloud);
 
 // Return coeff of y = c_0 * cos(x) + c_1 * sin(x)
-Eigen::Vector2f FitSine(const pcl::PointCloud<PointXYZPRVAE>::Ptr& cloud, const std::vector<int>& indices);
+Eigen::Vector2d FitSine(const pcl::PointCloud<PointXYZPRVAE>::Ptr& cloud, const std::vector<int>& indices);
 
 // Return Ransac outlier removed pointcloud
 pcl::PointCloud<PointXYZPRVAE>::Ptr RansacFit(const pcl::PointCloud<PointXYZPRVAE>::Ptr& cloud, float margin, int max_iterations);
